MINOR UPDATE
UPDATE
-Quando se apanha a comida especial o pacMan ganha Velocidade "tho"

+++ PAC-Man0.0.1/PAC-Man/Player.cs
@@ -15,27 +15,26 @@ namespace PAC_Man
             empowered
         };
         public PlayerState state;
-        public float speed;
+        public double speed;
 
         public Player(int Px, int Py)
         {
             this.Px = Px;
             this.Py = Py;
             this.state = PlayerState.normal;
-            this.speed = 5f;
-
+            this.speed = 5;
         }
         
-        public float Get_PacMan_Speed()
+        public double Get_PacMan_Speed()
         {
             return this.speed;
         }
 
         public void Change_Speed()
         {
-            if (speed == 5f)
-                speed = 2.5f;
-            else speed = 5f;
+            if (speed == 5)
+                speed = 2.5;
+            else speed = 5;
         }
 
         public void Set_Player_State()
